concatenation commit
We figured out how to get the label to have a multi-digit display and
got the value variable which keeps track of the display to reset each
time an operation key is pressed. We also changed test2 to be a
different test, one that we will probably end up using as our way of
adding numbers.